+++ packages/atomic/src/components/facets-v1/atomic-rating-facet/atomic-rating-facet.tsx
@@ -25,6 +25,7 @@ import {FacetValueLink} from '../facet-value-link/facet-value-link';
 import {FacetValueIconRating} from '../facet-value-icon-rating/facet-value-icon-rating';
 import {BaseFacet} from '../facet-common';
 import Star from '../../../images/star.svg';
+import {Schema, StringValue} from '@coveo/bueno';
 
 /**
  * A facet is a list of values for a certain field occurring in the results, ordered using a configurable criteria (e.g., number of occurrences).
@@ -91,6 +92,10 @@ export class AtomicRatingFacet
    */
   @Prop() public maxValueInIndex = this.numberOfIntervals;
   /**
+   * The minimum value of the field.
+   */
+  @Prop() public minValueInIndex = 1;
+  /**
    * Whether to display the facet values as checkboxes (multiple selection) or links (single selection).
    * Possible values are 'checkbox' and 'link'.
    */
@@ -100,7 +105,16 @@ export class AtomicRatingFacet
    */
   @Prop() public icon = Star;
 
+  private validateProps() {
+    new Schema({
+      displayValuesAs: new StringValue({constrainTo: ['checkbox', 'link']}),
+    }).validate({
+      displayValuesAs: this.displayValuesAs,
+    });
+  }
+
   public initialize() {
+    this.validateProps();
     this.searchStatus = buildSearchStatus(this.bindings.engine);
     this.initializeFacet();
   }
@@ -133,12 +147,12 @@ export class AtomicRatingFacet
 
   private generateCurrentValues() {
     const currentValues: NumericRangeRequest[] = [];
-    for (let i = 0; i < this.numberOfIntervals; i++) {
+    for (let i = this.minValueInIndex; i <= this.numberOfIntervals; i++) {
       currentValues.push(
         buildNumericRange({
           start: Math.round(i * this.scaleFactor * 100) / 100,
           end: Math.round((i + 1) * this.scaleFactor * 100) / 100,
-          endInclusive: true,
+          endInclusive: false,
         })
       );
     }
@@ -180,7 +194,7 @@ export class AtomicRatingFacet
           >
             <FacetValueIconRating
               numberOfTotalIcons={this.maxValueInIndex}
-              numberOfActiveIcons={facetValue.end}
+              numberOfActiveIcons={facetValue.start}
               icon={this.icon}
             ></FacetValueIconRating>
           </FacetValueCheckbox>
@@ -196,7 +210,7 @@ export class AtomicRatingFacet
           >
             <FacetValueIconRating
               numberOfTotalIcons={this.maxValueInIndex}
-              numberOfActiveIcons={facetValue.end}
+              numberOfActiveIcons={facetValue.start}
               icon={this.icon}
             ></FacetValueIconRating>
           </FacetValueLink>