+++ packages/atomic/src/components/facets-v1/facet-value-label-highlight/facet-value-label-highlight.tsx
@@ -1,6 +1,11 @@
 import {FunctionalComponent, h} from '@stencil/core';
-import {FacetValueLabelHighlightProps} from '../facet-common';
 import {highlightSearchResult} from '../facet-search/facet-search-utils';
+
+interface FacetValueLabelHighlightProps {
+  displayValue: string;
+  searchQuery?: string;
+  isSelected: boolean;
+}
 
 export const FacetValueLabelHighlight: FunctionalComponent<FacetValueLabelHighlightProps> = (
   props
@@ -9,7 +14,9 @@ export const FacetValueLabelHighlight: FunctionalComponent<FacetValueLabelHighli
     <span
       title={props.displayValue}
       part="value-label"
-      class={`value-label ellipsed ${props.isSelected ? 'font-bold' : ''}`}
+      class={`value-label w-full ellipsed ${
+        props.isSelected ? 'font-bold' : ''
+      }`}
       innerHTML={highlightSearchResult(props.displayValue, props.searchQuery)}
     ></span>
   );