+++ packages/quantic/force-app/main/default/lwc/quanticSort/quanticSort.js
@@ -1,5 +1,8 @@
 import {LightningElement, track, api} from 'lwc';
-import {registerComponentForInit, initializeWithHeadless} from 'c/quanticHeadlessLoader';
+import {
+  registerComponentForInit,
+  initializeWithHeadless,
+} from 'c/quanticHeadlessLoader';
 
 import sortBy from '@salesforce/label/c.quantic_SortBy';
 import relevancy from '@salesforce/label/c.quantic_Relevancy';
@@ -14,15 +17,25 @@ export default class QuanticSort extends LightningElement {
 
   /** @type {import("coveo").Sort} */
   sort;
-  /** @type {import("coveo").Unsubscribe} */
-  unsubscribe;
+  /** @type {import("coveo").SearchStatus} */
+  searchStatus;
+
+  /** @type {() => void} */
+  unsubscribeSort;
+  /** @type {() => void} */
+  unsubscribeSearchStatus;
+
+  /**
+   * @type {boolean}
+   */
+  @track hasResults;
 
   labels = {
     sortBy,
     relevancy,
     newest,
-    oldest
-  }
+    oldest,
+  };
 
   connectedCallback() {
     registerComponentForInit(this, this.engineId);
@@ -38,11 +51,25 @@ export default class QuanticSort extends LightningElement {
   @api
   initialize(engine) {
     this.sort = CoveoHeadless.buildSort(engine);
-    this.unsubscribe = this.sort.subscribe(() => this.updateState());
+    this.searchStatus = CoveoHeadless.buildSearchStatus(engine);
+    this.unsubscribeSort = this.sort.subscribe(() => this.updateState());
+    this.unsubscribeSearchStatus = this.searchStatus.subscribe(() =>
+      this.updateState()
+    );
+  }
+
+  disconnectedCallback() {
+    if (this.unsubscribeSearchStatus) {
+      this.unsubscribeSearchStatus();
+    }
+    if (this.unsubscribeSort) {
+      this.unsubscribeSort();
+    }
   }
 
   updateState() {
     this.state = this.sort.state;
+    this.hasResults = this.searchStatus.state.hasResults;
   }
 
   /**
@@ -74,15 +101,22 @@ export default class QuanticSort extends LightningElement {
   }
 
   get dateDescending() {
-    return CoveoHeadless.buildDateSortCriterion(CoveoHeadless.SortOrder.Descending)
+    return CoveoHeadless.buildDateSortCriterion(
+      CoveoHeadless.SortOrder.Descending
+    );
   }
-  
+
   get dateAscending() {
-    return CoveoHeadless.buildDateSortCriterion(CoveoHeadless.SortOrder.Ascending)
+    return CoveoHeadless.buildDateSortCriterion(
+      CoveoHeadless.SortOrder.Ascending
+    );
   }
 
   get largest() {
-    return CoveoHeadless.buildFieldSortCriterion('size', CoveoHeadless.SortOrder.Descending)
+    return CoveoHeadless.buildFieldSortCriterion(
+      'size',
+      CoveoHeadless.SortOrder.Descending
+    );
   }
 
   get options() {
@@ -94,7 +128,7 @@ export default class QuanticSort extends LightningElement {
   }
 
   get value() {
-    if (!this.sort) {
+    if (!this.sort || !this.state.sortCriteria.expression) {
       return 'relevancy';
     }
     return this.state.sortCriteria.expression;