+++ packages/samples/headless-react/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.36.20](https://github.com/coveo/ui-kit/compare/headless-react@0.36.19...headless-react@0.36.20) (2021-08-02)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.36.19](https://github.com/coveo/ui-kit/compare/headless-react@0.36.18...headless-react@0.36.19) (2021-07-30)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.36.18](https://github.com/coveo/ui-kit/compare/headless-react@0.36.17...headless-react@0.36.18) (2021-07-29)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.36.17](https://github.com/coveo/ui-kit/compare/headless-react@0.36.16...headless-react@0.36.17) (2021-07-28)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.36.16](https://github.com/coveo/ui-kit/compare/headless-react@0.36.15...headless-react@0.36.16) (2021-07-28)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.36.15](https://github.com/coveo/ui-kit/compare/headless-react@0.36.14...headless-react@0.36.15) (2021-07-28)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.36.14](https://github.com/coveo/ui-kit/compare/headless-react@0.36.13...headless-react@0.36.14) (2021-07-28)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
 ## [0.36.13](https://github.com/coveo/ui-kit/compare/headless-react@0.36.12...headless-react@0.36.13) (2021-07-26)
 
 **Note:** Version bump only for package headless-react